Updates power to take a double as a base

+++ maths.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 using namespace std;
-void power(int base, int exponent)
+void power(double base, int exponent)
 {
-    int retVal = base;
+    double retVal = base;
     for (int i = 0; i < exponent-1; i++)
     {
         retVal = retVal * base;
@@ -13,6 +13,6 @@ void power(int base, int exponent)
 int main()
 {
     cout << "GitHub Test\n";
-    power(2, 3);
+    power(2.0, 3);
 }
